Rename test case and fix code style issues
