+++ src/core/etl/tests/Flow/ETL/Tests/Unit/Extractor/ChunkFlowTest.php
@@ -6,8 +6,9 @@ namespace Flow\ETL\Tests\Unit\Extractor;
 
 use Flow\ETL\Extractor\ChunkExtractor;
 use Flow\ETL\Tests\Double\FakeExtractor;
+use Flow\ETL\Tests\FlowTestCase;
 
-final class ChunkExtractorTest extends ExtractorTestCase
+final class ChunkFlowTest extends FlowTestCase
 {
     public function test_chunk_extractor() : void
     {
